refactor: 불필요한 원시값 포장 제거
