+++ src/test/kotlin/omock/StoneTest.kt
@@ -1,18 +1,27 @@
 package omock
 
 import omock.model.Column
-import omock.model.Coordinate
-import omock.model.Row
 import omock.model.Stone
-import org.assertj.core.api.Assertions
+import omock.model.Row
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 
 class StoneTest {
     @Test
-    fun `돌은 좌표를 가지고 있다`() {
-        val coordinate = Coordinate(row = Row("1"), column = Column("A"))
-        val stone = Stone(coordinate)
-        Assertions.assertThat(stone.coordinate).isEqualTo(coordinate)
+    fun `돌은 row, column를 가지고 있다`() {
+        val stone = Stone(row = Row("1"), column = Column("A"))
+
+        assertThat(stone.row.comma).isEqualTo("1")
+        assertThat(stone.column.comma).isEqualTo("A")
+    }
+
+    @Test
+    fun `row, column을 통해 돌을 재활용 할 수 있다`() {
+        val row = Row("1")
+        val column = Column("A")
+        val actual = Stone.from(row = row, column = column)
+        assertThat(actual.row).isEqualTo(row)
+        assertThat(actual.column).isEqualTo(column)
     }
 }
